requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.0
typer[all]>=0.9.0
pytest>=7.0.0
responses>=0.25.0
